Big refactor, support self-install and path setup on Windows
