+++ Directory.Build.props
@@ -1,0 +1,9 @@
+<Project>
+    <PropertyGroup>
+        <RepoRoot>$([System.IO.Path]::GetFullPath('$(MSBuildThisFileDirectory)'))</RepoRoot>
+        <ArtifactsPath>$(RepoRoot)artifacts/</ArtifactsPath>
+        <BaseOutputPath>$(ArtifactsPath)bin/$(MSBuildProjectName)/</BaseOutputPath>
+        <BaseIntermediateOutputPath>$(ArtifactsPath)obj/$(MSBuildProjectName)/</BaseIntermediateOutputPath>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+</Project>
